pyproject.toml
[project]
name = "agentic-cyber-defense"
version = "0.1.0"
description = "Policy-aware multi-agent cybersecurity defense system"
requires-python = ">=3.10,<3.14"

dependencies = [
    "python-dotenv",
    "google-genai>=0.1.0",
    "numpy>=2.2.6",
    "scikit-learn>=1.7.2",
    "tensorflow-cpu>=2.20.0; python_version < '3.14'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest",
    "black",
    "ruff",
    "mypy"
]

[tool.hatch.build.targets.wheel]
packages = [
    "src/agents",
    "src/inference",
    "src/llm",
    "src/log_parser",
]
